sri/auth_system/backend/requirements.txt
fastapi
uvicorn
passlib[argon2]
python-jose[cryptography]
python-multipart
pydantic
